Add ability to have multiple errors per event
When implementing the Java library it was recognised that a single event
may actually have more than one underlying error. In light of this the
Java implementation opted to change the singular `error` field to an
`errors` array.

This change to the spec is to align to that change and to move towards
consistency between all of our implementations.